feat: modified authorizedLayout styles & header
