+++ src/shared/ui/appLogo/AppLogo.module.scss
@@ -1,0 +1,21 @@
+.logoWrapper {
+  display: flex;
+  align-items: center;
+  margin: 6px 0;
+}
+
+.logo {
+  display: flex;
+  justify-content: flex-start;
+  width: 200px;
+  height: auto;
+}
+
+.logo path {
+  fill: var(--accent-500);
+  transition: fill 0.3s;
+}
+
+.logo:hover path {
+  fill: var(--success-300);
+}
