+++ src/shared/ui/authButtons/AuthButtons.module.scss
@@ -1,0 +1,3 @@
+.item {
+  text-decoration: none;
+}
