+++ src/widget/header/Header.module.scss~
@@ -1,4 +1,8 @@
 .header {
+  position: sticky;
+  top: 0;
+  z-index: 100;
+
   display: flex;
   justify-content: center;
 
@@ -13,13 +17,6 @@
   display: flex;
   justify-content: space-between;
   width: 100%;
-  margin: 0 60px;
-}
-
-.logoWrapper {
-  display: flex;
-  align-items: center;
-  margin: 6px 0;
 }
 
 .itemsWrapper {
@@ -29,22 +26,6 @@
   justify-content: flex-end;
 
   width: 100%;
-}
-
-.logo {
-  display: flex;
-  justify-content: flex-start;
-  width: 200px;
-  height: auto;
-}
-
-.logo path {
-  fill: var(--accent-500);
-  transition: fill 0.3s;
-}
-
-.logo:hover path {
-  fill: var(--success-300);
 }
 
 s.bellWrapper {
@@ -57,57 +38,4 @@ s.bellWrapper {
 .bell {
   width: 24px;
   height: 24px;
-}
-
-.item {
-  text-decoration: none;
-}
-
-/*
-*_________________BELL____________________________
-*/
-
-@import '@/myApp/styles/mixins';
-
-.DropdownNotifications {
-  @include animation-appearance;
-
-  display: flex;
-  flex-flow: column nowrap;
-  gap: 15px;
-  align-items: flex-start;
-  justify-content: center;
-
-  padding: 10px;
-
-  background-color: var(--dark-700);
-  border: 1px solid var(--dark-300);
-  border-radius: 6px;
-}
-
-.notificationItemWrap {
-  display: flex;
-  flex-direction: column;
-
-  padding-right: 0;
-  padding-bottom: 0;
-  padding-left: 0;
-}
-
-.notificationEl {
-  display: flex;
-  flex-direction: column;
-  align-items: center;
-  justify-content: center;
-
-  width: 100%;
-  
-  &:not(:last-child) {
-    padding-bottom: 12px;
-    border-bottom: 1px solid var(--dark-300);
-  }
-}
-
-.notificationItem {
-   /* border: 1px solid red; */
 }
